+++ rockpool/devices/xylo/syns65302/afe/__init__.py
@@ -12,4 +12,4 @@ from .digital_filterbank import ChipButterworth
 from .raster import Raster
 from .divisive_normalization import DivisiveNormalization
 from .params import *
-from .external import *
+from .resample_and_quantize import *