+++ rockpool/devices/xylo/syns65302/afe_samna_pdm.py
@@ -1,5 +1,5 @@
 """
-Provides :py:class:`.XyloSamnaPDM`
+Provides :py:class:`.AFESamnaPDM`
 """
 
 import numpy as np
@@ -31,7 +31,7 @@ XyloAudio3HDK = samna.xyloAudio3.XyloAudio3TestBoard
 __all__ = ["XyloSamnaPDM"]
 
 
-class XyloSamnaPDM(Module):
+class AFESamnaPDM(Module):
     """
     A spiking neuron :py:class:`.Module` backed by the Xylo hardware, via `samna`, with PDM input.
     """
@@ -39,7 +39,7 @@ class XyloSamnaPDM(Module):
     def __init__(
         self,
         device: XyloAudio3HDK,
-        snn_config: XyloConfiguration = None,
+        config: XyloConfiguration = None,
         # pdm_config: PdmPreprocessingConfig = None,
         # dfe_config: DigitalFrontendConfig = None,
         dt: float = 1024e-6,
@@ -50,20 +50,24 @@ class XyloSamnaPDM(Module):
         **kwargs,
     ):
         """
-        Instantiate a Module with PdmEvents as input source for a XyloAudio 3 dev-kit backend.
+        Instantiate a Module with Xylo dev-kit backend
+        This module uses PdmEvents as input source. It bypass the microphone, feeding the input data (PdmEvents) to the filter bank.
+        The output responses are both the generated input spikes that will be fed into the SNN core and the result of the SNN core processing.
 
         Args:
-            device (XyloAudio3HDK): An opened `samna` device to a XyloAudio 3 dev kit.
-            config (XyloConfiguration): A Xylo configuration from `samna`.
-            dt (float): The simulation time-step to use for this Module. Default: 1024e-6.
+            device (XyloAudio3HDK): An opened `samna` device to a XyloAudio 3 dev kit
+            config (XyloConfiguration): A Xylo configuration from `samna`
+            dt (float): The simulation time-step to use for this Module
             output_mode (str): The readout mode for the Xylo device. This must be one of ``["Spike", "Isyn", "Vmem"]``. Default: "Spike", return events from the output layer.
-            power_frequency (float): The frequency of power measurement in Hz. Default: 100 Hz.
+            record (bool): Record and return all internal state of the neurons and synapses on Xylo. Default: ``False``, do not record internal state.
+            power_frequency (float): The frequency of power measurement. Default: 5.0
             dn_active (bool): If True, divisive normalization will be used. Defaults to True.
 
         Raises:
             `ValueError`: If ``device`` is not set. ``device`` must be a ``XyloAudio3HDK``.
-            `ValueError`: If ``output_mode`` is not ``Spike``, ``Vmem`` or ``ISyn``.
-            `ValueError`: If `snn_config.operation_mode` is set to `RealTime`. Valid options are `AcceleratedTime` and `Manual`.
+            `ValueError`: If ``output_mode`` is not ``Spike``, ``Vmem`` or ``Isyn``.
+            `ValueError`: If ``operation_mode`` is ``RealTime``. Valid options are ``Manual`` or ``AcceleratedTime``.
+
         """
 
         # - Check input arguments
@@ -78,17 +82,17 @@ class XyloSamnaPDM(Module):
         self._output_mode = output_mode
 
         # - Get a default configuration
-        if snn_config is None:
-            snn_config = samna.xyloAudio3.configuration.XyloConfiguration()
+        if config is None:
+            config = samna.xyloAudio3.configuration.XyloConfiguration()
 
-        # - Set input source to PdmEvents
-        snn_config.input_source = samna.xyloAudio3.InputSource.PdmEvents
+        # - Input source must be PdmEvents
+        config.input_source = samna.xyloAudio3.InputSource.PdmEvents
 
         # - Get the network shape
-        Nin, _ = np.shape(snn_config.input.weights)
-        Nhidden, _ = np.shape(snn_config.hidden.weights)
+        Nin, _ = np.shape(config.input.weights)
+        Nhidden, _ = np.shape(config.hidden.weights)
 
-        _, Nout = np.shape(snn_config.readout.weights)
+        _, Nout = np.shape(config.readout.weights)
 
         # - Initialise the superclass
         super().__init__(
@@ -115,28 +119,22 @@ class XyloSamnaPDM(Module):
 
         # - Configure the PDM usage
         # TODO FIXME: Allow hibnernation mode to be selectable - https://www.wrike.com/open.htm?id=1552655948
-        snn_config.digital_frontend.filter_bank.dn_enable = dn_active
-        snn_config.digital_frontend.hibernation_mode_enable = 0
-        snn_config.digital_frontend.filter_bank.use_global_iaf_threshold = 1
-        snn_config.digital_frontend.pdm_preprocessing.clock_direction = 0
-        snn_config.digital_frontend.pdm_preprocessing.clock_edge = 0
+        config.digital_frontend.filter_bank.dn_enable = dn_active
+        config.digital_frontend.hibernation_mode_enable = 0
+        config.digital_frontend.filter_bank.use_global_iaf_threshold = 1
+        config.digital_frontend.pdm_preprocessing.clock_direction = 0
+        config.digital_frontend.pdm_preprocessing.clock_edge = 0
 
-        if snn_config.operation_mode == samna.xyloAudio3.OperationMode.AcceleratedTime:
-            snn_config.debug.always_update_omp_stat = True
+        if config.operation_mode == samna.xyloAudio3.OperationMode.AcceleratedTime:
+            config.debug.always_update_omp_stat = True
 
-        # - Store the SNN core configuration (and apply it)
         time.sleep(self._sleep_time)
 
-        # - For XyloSamnaPDM, operation mode can be either manual or accelerated time
-        if snn_config.operation_mode == samna.xyloAudio3.OperationMode.RealTime:
+        # - For AFESamnaPDM, operation mode can be either manual or accelerated time
+        if config.operation_mode == samna.xyloAudio3.OperationMode.RealTime:
             raise ValueError(
-                "`operation_mode` can't be RealTime for XyloSamnaPDM. Options are Manual or AcceleratedTime."
+                "`operation_mode` can't be RealTime for AFESamnaPDM. Options are Manual or AcceleratedTime."
             )
-
-        self._snn_config: Union[
-            XyloConfiguration, SimulationParameter
-        ] = SimulationParameter(shape=(), init_func=lambda _: snn_config)
-        """ `.XyloConfiguration`: The HDK configuration applied to the Xylo module """
 
         # - Store the power frequency
         self._power_frequency = power_frequency
@@ -157,8 +155,14 @@ class XyloSamnaPDM(Module):
         self._last_record_mode: Optional[bool] = None
         """ bool: The most recent (and assumed still valid) recording mode """
 
+        # - Store the SNN core configuration
+        self._config: Union[
+            XyloConfiguration, SimulationParameter
+        ] = SimulationParameter(shape=(), init_func=lambda _: config)
+        """ `.XyloConfiguration`: The HDK configuration applied to the Xylo module """
+
         # - Apply configuration on the board
-        hdkutils.apply_configuration(self._device, self._snn_config)
+        hdkutils.apply_configuration(self._device, self._config)
 
     @property
     def config(self):
@@ -188,8 +192,7 @@ class XyloSamnaPDM(Module):
         # - Reset the HDK to clean up
         self._device.reset_board_soft()
 
-    def reset_state(self) -> "XyloSamnaPDM":
-        # - Reset neuron and synapse state on Xylo
+    def reset_state(self) -> "AFESamnaPDM":
         # - Reset neuron and synapse state on Xylo
         # TODO FIXME - https://www.wrike.com/open.htm?id=1533940426 - reset state is not working
         warn("Reset state is not working yet.")
@@ -201,6 +204,7 @@ class XyloSamnaPDM(Module):
         record: bool = False,
         record_power: bool = False,
         read_timeout: float = 5.0,
+        flip_and_encode: bool = False,
         *args,
         **kwargs,
     ) -> Tuple[np.ndarray, dict, dict]:
@@ -214,7 +218,8 @@ class XyloSamnaPDM(Module):
             input (np.ndarray): A vector ``(Tpdm, 1)`` with a PDM-encoded audio signal, ``1`` or ``0``. The PDM clock is always 1.5625 MHz. 32 PDM samples correspond to one audio sample passed to the band-pass filterbank (i.e. 48.828125 kHz). The network ``dt`` is independent of this sampling rate, but should be an even divisor of 48.828125 MHz (e.g. 1024 us).
             record (bool): Record and return all internal state of the neurons and synapses on Xylo. Default: ``False``, do not record internal state.
             record_power (bool): Iff ``True``, record the power consumption during each evolve.
-            read_timeout (Optional[float]): Set an explicit read timeout for the entire simulation time. This should be sufficient for the simulation to complete, and for data to be returned. Default: ``None``, set a reasonable default timeout.
+            read_timeout (Optional[float]): Set an explicit read timeout for the entire simulation time. This should be sufficient for the simulation to complete, and for data to be returned. Default: 5s.
+            flip_and_encode (bool): Determine if flip-and-encode should be applied to the input data. When applied, the input data will be flipped on axis=0 and concatenated to the begin of the original input data. Note that input data will have its size doubled.
 
         Returns:
             (np.ndarray, dict, dict): ``output``, ``new_state``, ``record_dict``.
@@ -223,7 +228,7 @@ class XyloSamnaPDM(Module):
             ``record_dict`` is a dictionary containing recorded internal state of Xylo during evolution, if the ``record`` argument is ``True``. Otherwise this is an empty dictionary.
 
         Raises:
-            `ValueError`: If `snn_config.operation_mode` is set to `RealTime`. Valid options are `AcceleratedTime` and `Manual`.
+            `ValueError`: If ``operation_mode`` is ``RealTime``. Valid options are ``Manual`` or ``AcceleratedTime``.
             `TimeoutError`: If reading data times out during the evolution. An explicity timeout can be set using the `read_timeout` argument.
         """
 
@@ -231,9 +236,9 @@ class XyloSamnaPDM(Module):
         Nin, Nhidden, Nout = self.shape
 
         # - Check again if operation mode is either manual or accelerated time
-        if self._snn_config.operation_mode == samna.xyloAudio3.OperationMode.RealTime:
+        if self._config.operation_mode == samna.xyloAudio3.OperationMode.RealTime:
             raise ValueError(
-                "`operation_mode` can't be RealTime for XyloSamnaPDM. Options are Manual or AcceleratedTime."
+                "`operation_mode` can't be RealTime for AFESamnaPDM. Options are Manual or AcceleratedTime."
             )
 
         if record != self._last_record_mode:
@@ -275,6 +280,13 @@ class XyloSamnaPDM(Module):
             input[: num_dt * PDM_samples_per_dt], [-1, PDM_samples_per_dt]
         )
 
+        flip_and_encode_size = None
+        if flip_and_encode:
+            # -- Revert and repeat the input signal in the beginning to avoid boundary effects
+            flip_and_encode_size = np.shape(input_raster)[0]
+            __input_rev = np.flip(input_raster, axis=0)
+            input_raster = np.concatenate((__input_rev, input_raster), axis=0)
+
         # - Initialise lists for recording state
         input_spikes = []
         vmem_ts = []
@@ -304,7 +316,7 @@ class XyloSamnaPDM(Module):
                 )
 
             # - Trigger processing: if manual mode, advance time step manually
-            if self._snn_config.operation_mode == samna.xyloAudio3.OperationMode.Manual:
+            if self._config.operation_mode == samna.xyloAudio3.OperationMode.Manual:
                 hdkutils.advance_time_step(self._write_buffer)
                 # - Wait until xylo has finished the simulation of this time step
                 t_start = time.time()
@@ -376,6 +388,14 @@ class XyloSamnaPDM(Module):
                     "digital_power": digital_power,
                 }
             )
+        output_ts = np.array(output_ts)
+
+        if flip_and_encode:
+            # Trim the part of the signal coresponding to __input_rev (which was added to avoid boundary effects)
+            output_ts = output_ts[flip_and_encode_size:, :]
+
+            # # Trim recordings
+            rec_dict = {k: v[flip_and_encode_size:, :] for k, v in rec_dict.items()}
 
         # - Return the output spikes, the (empty) new state dictionary, and the recorded state dictionary
-        return np.array(output_ts), {}, rec_dict
+        return output_ts, {}, rec_dict