+++ CHANGELOG.md
@@ -2,14 +2,16 @@
 
 All notable changes between Rockpool releases will be documented in this file
 
-## [v2.10.0.dev2] -- XXXX-XX-XX
+## [v2.10.0.dev3] -- XXXX-XX-XX
 
 ### Added
 * Power measurement for XyloAudio 3
+* AFESamna module for XyloAudio 3: allow recording of input spikes from live microphone
 
 ### Changed
 ### Fixed
 * Remove dependence of imp module that was deprecated in Python 3.12
+* Fixed bug in multiplying `TSContinuous` objects with differing numbers of channels
 
 ### Deprecated
 ### Removed